[FEATURE] Move JavaScript with Ajax request to external file
By using the 'FooterAssets' section in List.html, the JavaScript
is loaded from a new external file. It will be rendered just before
the closing </body> tag.

This allows to e.g. set a strict Content Security Policy (no inline
JavaScript) in a website where this extension is used.

This shouldn't be a breaking change. If you adjusted the List.html
template, you will continue to use the JavaScript provided there.
And if you didn't change it, the extension will load the new
external file.