[FEATURE] Allow to choose between jQuery.ajax() and XMLHttpRequest()
Until now, the extension only shipped with a jQuery solution for the
Ajax request.

The new TypoScript setting 'useAjaxwithJQuery' allows to switch to
a second solution, which uses an XMLHttpRequest (vanilla JavaScript).
By default, this is set to '1' (for backward compatibility).

This setting is considered in the 'List.html' template, where the
appropriate JavaScript file is rendered just before the closing
</body> tag.

The current state of the XMLHttpRequest is compatible with modern
browsers (no Internet Explorer).